backend/rating/requirements.txt
redis-om==0.0.27
sanic==22.6.2